fix(types): add casts and type assertions for torch/numpy interop
Category 4/6: Cast/conversion type fixes
- cuda_optimization.py: type ignore for torch.compile return
- unified_tools.py: explicit tensor initialization for merged sum
- vl_jepa_extension.py: class-level type hints for registered buffers
- interconnect_manager.py: cast np.mean to float before int conversion
- active_memory.py: type assertions for quantization_levels buffer

Fixes ~10 mypy type conversion errors

+++ src/cogsyndelta/core/interconnect_manager.py
@@ -402,7 +402,7 @@ class CongestionController:
             return 5  # Default priority
 
         # Higher usage = higher priority
-        avg_usage = np.mean(history[-10:])
+        avg_usage = float(np.mean(history[-10:]))
         return int(min(10, avg_usage / 100))
 
     def get_statistics(self) -> dict[str, Any]:
